+++ src/fetch.rs
@@ -800,7 +800,12 @@ fn print_available_streams_representation(r: &Representation, a: &AdaptationSet,
     } else {
         format!("{w}x{h}")
     };
-    println!("  {typ} {codec:17} | {:5} Kbps | {fmt:>9}", bw / 1024);
+    let role = a.Role.iter().next()
+        .map_or_else(|| String::from(""),
+                     |r| r.value.as_ref().map_or_else(|| String::from(""), |v| format!(" role={v}")));
+    let label = a.Label.iter().next()
+        .map_or_else(|| String::from(""), |l| format!(" label={}", l.clone().content));
+    println!("  {typ} {codec:17} | {:5} Kbps | {fmt:>9}{role}{label}", bw / 1024);
 }
 
 fn print_available_streams_adaptation(a: &AdaptationSet, typ: &str) {